// src/models/enum/Colors.ts
export enum ThemeColors {
  Primary = "primary",
  Primary100 = "primary-100",
  Primary200 = "primary-200",
  Primary300 = "primary-300",
  Primary400 = "primary-400",

  Secondary = "secondary",
}

export enum GraysColors {
  White = "white",
  Gray100 = "gray-100",
  Gray200 = "gray-200",
  Gray300 = "gray-300",
  Gray400 = "gray-400",
  Gray500 = "gray-500",
  Gray600 = "gray-600",
  Gray700 = "gray-700",
  Gray800 = "gray-800",
  Gray900 = "gray-900",
  Gray = "gray",
}

export enum RedColors {
  Red = "red",
  Red100 = "red-100",
  Red200 = "red-200",
  Red300 = "red-300",
  Red400 = "red-400",
  Red500 = "red-500",
}

export enum YellowColors {
  Yellow100 = "yellow-100",
  Yellow200 = "yellow-200",
  Yellow300 = "yellow-300",
  Yellow400 = "yellow-400",
  Yellow500 = "yellow-500",
  Yellow600 = "yellow-600",
  Yellow700 = "yellow-700",
  Yellow800 = "yellow-800",
  Yellow900 = "yellow-900",
}

export enum GreenColors {
  Green100 = "green-100",
  Green200 = "green-200",
  Green300 = "green-300",
  Green400 = "green-400",
  Green500 = "green-500",
  Green600 = "green-600",
}

export enum VioletColors {
  Violet100 = "violet-100",
  Violet200 = "violet-200",
  Violet300 = "violet-300",
  Violet400 = "violet-400",
  Violet500 = "violet-500",
  Violet600 = "violet-600",
  Violet700 = "violet-700",
  Violet800 = "violet-800",
  Violet900 = "violet-900",
}